Fix maximum absolute effect size calculation
The initial value should be 0 instead of f32::MIN because
f32::MIN.abs() is a very large positive number, much larger than
any of the effect sizes, so it will always be the final result.

Also, while we need to compare the absolute values of effect sizes
to get the largest effect size we should save the non-absolute value.
Otherwise we'll never show negative effect sizes.